requirements.txt
# Core dependencies
numpy>=1.24.0
opencv-python>=4.7.0
PyGObject>=3.42.0  # For GStreamer Python bindings
PyGI>=1.0.0

# AI/ML (CPU-only versions)
tensorflow-cpu>=2.8.0
torch>=2.0.0
torchvision>=0.15.0
pandas>=1.3.0
seaborn>=0.12.0
--extra-index-url https://download.pytorch.org/whl/cpu

# REST API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# MQTT
paho-mqtt==1.6.1

# Utilities
python-dotenv==1.0.0
loguru>=0.7.0

# Optional: Time series database
influxdb-client>=1.36.0
matplotlib>=3.7.0

# Additional dependencies
python-multipart==0.0.6
psutil>=5.9.0
requests==2.31.0
aiohttp>=3.8.0
pyyaml>=6.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pytest==7.4.3
ultralytics==8.0.196 